+++ NanoXLSX.Core/Changelog.md
@@ -3,9 +3,12 @@
 ## v3.0.0-rc.3
 
 ---
-Release Date: **11.12.2025**
+Release Date: **04.01.2026**
 
 - Internal change of structured text handling
+- Formal change of the `Color` and `ThemeColor` classes
+- Removed the property `ColorTheme` from the `Font` class
+- Changed the type of the property `ColorVlaue` of the Font class from `string` to `Color` (namespace `NanoXLES.Colors`)
 
 
 ## v3.0.0-rc.2